aliases: Support machine specific files & folders
The alias management flow has been improved to allow
custom folders and "configs" (should be named "files" more
generally later) for specific computers. To use this, place the
desired files and folder locations into the file named
configs-<hostname> or folders-<hostname>, where <hostname> is
the hostname of that machine with those specific aliases.

The shortcuts.sh script has been cleaned up to extend to these
new machine-specific files, and it will stitch them together in the
output autoaliases files. Due to this, these autoaliases files will
no longer be included on the repo directly (it was also an extra
annoyance when managing changes to them in git).

Vim has been re-configured as well to automatically run the
shortcuts.sh script when one of these custom machine-specific alias
files is modified.