bash/DESKTOP-DNH8H8Q: Force working directory
When using Alacritty under Windows with WSL, things are getting
a little more strange, and I'm not 100% sure how to make this thing
open at the right directory. So as a work-around, I'm setting bash
to change to the home directory on start. It's not a perfect
solution, but it does well enough.