bashrc: Hide output if something wrong happened
Add some more checks in place so that now, if this script was running
on a machine that didn't have a certain file or program, it doesn't
spill a bunch of errors all over the terminal. Not sure if this is
all-inclusive right now, but it's better than it was yesterday.